snapcraft: make it clear that lz4 lib isn't primed
snapcraft deduplication is not applied to files from packages listed in
stage-packages.  To avoid duplicating lz4's .so, make it explicit that we don't
want to prime it and instead relying on the one provided in the base snap.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
(cherry picked from commit ccaef86bc3d1f791068006b8efd745ad49f01d01)